Enhancements to shift change reminders via text messages
NOTE: Prior to using this changeset, someone must navigate to the
database_update.php page, since a new column has been added to the
call_times database table (see below).

Removed code that tried to be clever about pruning out shift change
messages that are duplicates of one another (e.g. person X has a
shift for texts and calls starting at 12:10am). The idea was to avoid
annoying duplicates, but it was pointed out at the last tech team
meeting that duplicate messages mean duplicate entries in the database,
which should be cleaned up, so staff members will now see such
duplicates if shifts that have parameters in common exist.

Added the type(s) of relevant shift (texts, calls, and/or call alert
messages) to the notification sent via text to the staff member, so
that the latter will know what they're signed up to handle.

Added weeding out of notifications that should not be sent out because
one shift is ending at 11:59:59pm and the next one, identical in terms
of the staff member's name and the types of incoming info they are to
handle, is starting at 12:00am the next day. These are used because
entries cannot span multiple days, so if for example a staff member is
on duty from 8:00pm to 4:00am, this has to be entered into the database
as two shifts, one lasting until midnight, the next starting at
midnight. Such cases will no longer result in the staff member being
told that they have  shift ending at 11:59pm and then another one
starting at 12:00am. (Note that if the type of incoming communications
handled is different between the two entries, then the staff member
will be notified, as those are two different shifts.)

Added a new column to the call_times table in the database, "remind",
that indicates whether or not a shift's start or end is to generate a
reminder text message. This defaults to true. The add/edit call time
modal dialog has been changed to include a checkbox for this parameter,
so it can be toggled for individual shifts easily.